+++ front-end/src/api/index.js
@@ -23,15 +23,6 @@ async function registerUser(userData) {
   }
 }
 
-async function updateUser(userData) {
-  try {
-    const response = await axios.put(`${url}${register}`, userData);
-    return response;
-  } catch (error) {
-    return error.response.status;
-  }
-}
-
 async function getProducts(token) {
   try {
     const productsReceived = axios
@@ -41,5 +32,17 @@ async function getProducts(token) {
     return error.response.status;
   }
 }
+
+
+async function updateUser(userData) {
+  console.log(userData);
+  try {
+    const response = await axios.put(`${url}${register}`, userData);
+    return response;
+  } catch (error) {
+    return error.response.status;
+  }
+}
+
 
 export { requestToken, registerUser, getProducts, updateUser };
