+++ front-end/src/components/ProfileForm.js
@@ -1,0 +1,89 @@
+import React, { useState } from 'react';
+import { Form, Button } from 'react-bulma-components';
+import services from '../services';
+import { updateUser } from '../api';
+
+const { Input, Field, Control, Label } = Form;
+
+function ProfileForm() {
+
+  const [inputName, setInputName] = useState('');
+  const [notification, setNotification] = useState(true);
+  const errorCode = 400;
+
+  const { email, name, role, token } = services.getUserLocalStorage();
+
+  const handleClick = async () => {
+    const newUser = {
+      email,
+      name: inputName,
+      role,      
+      token,
+    };
+    await updateUser({ name: inputName, email})
+    services.setUserLocalStorage(newUser); 
+    setNotification(false);        
+    }      
+  
+
+  return (
+    <>
+    <h1
+        data-testid="top-title"
+      >
+        Meu perfil
+      </h1>
+    <div
+        hidden={ notification }
+      >
+        <p>
+        Atualização concluída com sucesso
+          <Button
+            remove
+            onClick={ () => setNotification(true) }
+          />
+        </p>
+      </div>
+    <Field>
+        <Label>
+          Nome
+        </Label>
+        <Control>
+          <Input
+            placeholder={ name }
+            onChange={ (e) => setInputName(e.target.value) }
+            data-testid="profile-name-input"
+          />
+        </Control>
+      </Field>
+
+      <Field>
+        <Label>
+          Email
+        </Label>
+        <Control>
+          <Input
+            data-testid="profile-email-input"
+            type="text"
+            value={ email }
+            readOnly
+          />
+        </Control>
+      </Field>
+
+      <Field>
+        <Control>
+          <Button
+            onClick={ () => handleClick() }
+            data-testid="profile-save-btn"
+            disabled={ !inputName }
+          >
+            Salvar
+          </Button>
+        </Control>
+      </Field>
+    </>
+  );
+}
+
+export default ProfileForm;
