+++ front-end/src/components/index.js
@@ -2,10 +2,12 @@ import LoginForm from './LoginForm';
 import RegisterForm from './RegisterForm';
 import ProductCard from './ProductCard';
 import Loading from './Loading/Loading';
+import ProfileForm from './ProfileForm';
 
 export {
   LoginForm,
   RegisterForm,
   ProductCard,
   Loading,
+  ProfileForm,
 };
